Project Main/requirementx.txt
Flask==2.2.3
openai==0.27.2
mysql-connector-python==8.3.0
